requirements.txt
# TOSCA Laser Control System - Python Dependencies
# Python Version: 3.10+

# ===================================================================
# GUI Framework & Visualization
# ===================================================================
PyQt6>=6.6.0
PyQt6-Qt6>=6.6.0
pyqtgraph>=0.13.3

# ===================================================================
# Image Processing & Computer Vision
# ===================================================================
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# ===================================================================
# Hardware Interfaces
# ===================================================================
pyserial>=3.5
adafruit-blinka>=8.20.0
adafruit-circuitpython-busdevice>=5.2.0

# Note: Install these separately from vendor SDKs:
# - Xeryon actuator library (vendor-specific)
# - VmbPy (Allied Vision camera SDK)

# ===================================================================
# Database & ORM
# ===================================================================
sqlalchemy>=2.0.0
alembic>=1.12.0

# ===================================================================
# Data Validation & Configuration
# ===================================================================
pydantic>=2.0.0
pydantic-settings>=2.0.0
jsonschema>=4.19.0

# ===================================================================
# Logging & Utilities
# ===================================================================
python-dateutil>=2.8.2
loguru>=0.7.0

# ===================================================================
# Development & Testing
# ===================================================================
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-qt>=4.2.0
pytest-mock>=3.11.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
pylint>=2.17.0

# ===================================================================
# Type Stubs for Static Analysis
# ===================================================================
types-python-dateutil>=2.8.0

# ===================================================================
# Safety & Security Analysis
# ===================================================================
safety>=2.3.0
bandit>=1.7.5

# ===================================================================
# Documentation
# ===================================================================
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# ===================================================================
# Additional Utilities
# ===================================================================
python-dotenv>=1.0.0

# ===================================================================
# Camera Hardware
# ===================================================================
vmbpy>=1.1.0
